api/requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mongoengine==0.20.0
Pillow==7.1.2
praw==7.0.0
prawcore==1.3.0
pymongo==3.10.1
python-dotenv==0.13.0
requests==2.23.0
six==1.14.0
update-checker==0.17
urllib3==1.25.9
websocket-client==0.57.0
Werkzeug==1.0.1
